feat: Update icons and components
