+++ src/components/atoms/Icon/index.tsx
@@ -1,8 +1,6 @@
 import styled from '@mui/system/styled';
 
 interface IconProps {
-  id?: string | undefined;
-  className?: string | undefined;
   src: string;
   alt: string;
   width?: string | number;
@@ -17,8 +15,8 @@ const StyledIcon = styled('img')<IconProps>(({ width, height }) => ({
 
 
 // id and classnames are for future adaptions for external css.
-const Icon: React.FC<IconProps> = ({ id, className, src, alt, width = 'auto', height = 'auto' }) => {
-  return <StyledIcon id={id} className={className} src={src} alt={alt} width={width} height={height} />;
+const Icon: React.FC<IconProps> = ({ src, alt, width = 'auto', height = 'auto' }) => {
+  return <StyledIcon src={src} alt={alt} width={width} height={height} />;
 };
 
 export default Icon;
