+++ src/pages/CardPage/index.tsx
@@ -1,6 +1,5 @@
 import CardTemplate from "../../components/template/CardTemplate";
 
-
 const CardPage = () => {
     return <CardTemplate />;
 }
